+++ source/bind_scene.cpp
@@ -12,7 +12,7 @@ namespace Bind {
 
 void CheckAndInitPropertyTable(lua_State* pLua, Node* pNode) {
     lua_getfield(pLua, LUA_REGISTRYINDEX, "_nodePropTables");
-    lua_pushinteger(pLua, pNode->id);
+    lua_pushinteger(pLua, (i32)pNode->id);
     lua_gettable(pLua, -2);
 
     if (lua_type(pLua, -1) != LUA_TTABLE) {
@@ -28,7 +28,7 @@ void CheckAndInitPropertyTable(lua_State* pLua, Node* pNode) {
         lua_pushlstring(pLua, pNode->name.pData, pNode->name.length);
         lua_setfield(pLua, -2, "name");
 
-        lua_pushinteger(pLua, pNode->id);
+        lua_pushinteger(pLua, (i32)pNode->id);
         lua_pushvalue(pLua, -2);
         lua_settable(pLua, -4);  // Sets _nodePropTables[nodeId] = newTable (and leaves new table on the stack)
     }
@@ -180,14 +180,14 @@ int Node_GetPropertyTable(lua_State* pLua) {
     lua_getfield(pLua, LUA_REGISTRYINDEX, "_nodePropTables");
 
     // node key is an integer (either it's ptr, or eventually a bit merge of scene and node key)
-    lua_pushinteger(pLua, pNode->id);
+    lua_pushinteger(pLua, (i32)pNode->id);
     lua_gettable(pLua, -2);
 
     // We check if that element exists, if not create a new table like _nodePropTables[nodeKey] = t
     if (lua_type(pLua, -1) != LUA_TTABLE) {
         lua_pop(pLua, 1);
         lua_newtable(pLua);
-        lua_pushinteger(pLua, pNode->id);
+        lua_pushinteger(pLua, (i32)pNode->id);
         lua_pushvalue(pLua, -2);
         lua_settable(pLua, -4);  // Sets _nodePropTables[nodeId] = newTable (and leaves new table on the stack)
     }
